INTYGFV-15592: Added total number of sick leaves to populated filters response and refactored filtering on protected person
