+++ integration/it-integration/src/main/java/se/inera/intyg/rehabstod/integration/it/dto/PopulateFiltersResponseDTO.java
@@ -32,4 +32,5 @@ import se.inera.intyg.infra.sjukfall.dto.Lakare;
 public class PopulateFiltersResponseDTO {
     private List<Lakare> activeDoctors;
     private List<DiagnosKapitel> diagnosisChapters;
+    private int nbrOfSickLeaves;
 }
