+++ web/src/main/java/se/inera/intyg/rehabstod/service/sjukfall/dto/PopulateFiltersResponseDTO.java
@@ -17,17 +17,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package se.inera.intyg.rehabstod.integration.it.dto;
+package se.inera.intyg.rehabstod.service.sjukfall.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import se.inera.intyg.rehabstod.service.diagnos.dto.DiagnosKapitel;
+import se.inera.intyg.rehabstod.web.model.Lakare;
+
+import java.util.List;
 
 @Data
 @NoArgsConstructor
-public class PopulateFiltersRequestDTO {
-    private String unitId;
-    private String careUnitId;
-    private String doctorId;
-    private int maxDaysSinceSickLeaveCompleted;
-    private boolean filterProtectedPerson;
+@AllArgsConstructor
+public class PopulateFiltersResponseDTO {
+    private List<Lakare> activeDoctors;
+    private List<DiagnosKapitel> allDiagnosisChapters;
+    private List<DiagnosKapitel> enabledDiagnosisChapters;
+    private int nbrOfSickLeaves;
 }
