+++ web/src/main/java/se/inera/intyg/rehabstod/web/controller/api/dto/PopulateFiltersResponseDTO.java
@@ -33,4 +33,5 @@ public class PopulateFiltersResponseDTO {
     private List<Lakare> activeDoctors;
     private List<DiagnosKapitel> allDiagnosisChapters;
     private List<DiagnosKapitel> enabledDiagnosisChapters;
+    private int nbrOfSickLeaves;
 }
